+++ src/main/java/graph/Node.java
@@ -1,6 +1,6 @@
 package graph;
 
-public record Node(int osmid, float longitude, float latitude) {
+public record Node(long osmid, float longitude, float latitude) {
     public Node deepCopy() {
         return new Node(this.osmid, this.longitude, this.latitude);
     }
