+++ src/main/java/osmxml/WayContext.java
@@ -3,11 +3,11 @@ package osmxml;
 import java.util.ArrayList;
 
 public class WayContext {
-    public int id;
+    public Long id;
     public boolean isRoad;
     public String roadType;
-    public ArrayList<Integer> nodesInPath;
-    public WayContext(int id) {
+    public ArrayList<Long> nodesInPath;
+    public WayContext(Long id) {
         this.id = id;
         nodesInPath = new ArrayList<>();
     }
